power: supply: ug3105_battery: Add driver for uPI uG3105 battery monitor
Add a new battery driver for the uPI uG3105 battery monitor.

Note the uG3105 is not a full-featured autonomic fuel-gauge. Instead it is
expected to be use in combination with some always on microcontroller reading
its coulomb-counter before it can wrap (must be read every 400 seconds!).

Since Linux does not monitor coulomb-counter changes while the device is
off or suspended, the coulomb counter is not used atm.

So far this driver is only used on x86/ACPI (non devicetree) devs
(also note there is no of_match table). Therefor there is no devicetree
bindings documentation for this driver's "upi,rinternal-microohm" and
"upi,rsns-microohm" properties since these are not used in actual
devicetree files. Note the devicetree-bindings maintainers have requested
properties with no actual dt users to _not_ be added to the dt bindings.

The property names have been chosen so that they hopefully do not need
to be changed if/when devicetree enumeration support gets added later.
The "upi,rsns-microohm" mirrors "maxim,rsns-microohm" from the
"maxim,max17042" bindings and the other 2 property names follow the
standard battery bindings.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>