device property: Helper to match multiple connections
In some cases multiple connections with the same connection id
needs to be resolved from a fwnode graph.

One such example is when separate hardware is used for performing muxing
and/or orientation switching of the SuperSpeed and SBU lines in a USB-C
connector. In this case the connector needs to belong to a graph with
multiple matching remote endpoints, and the TypeC controller needs to be
able to resolve them both.

Add a new API that allows this kind of lookup.

Given that the match() callback returns an opaque reference to something
provided by the client it's not possible for the implementation to
release the returned object and as such it's not possible to handle
errors, which in turn means that it's not possible to query the number
of elements or dynamically grow the results array. It's however expected
that the number of matches will be reasonably low and that the worst
case is known by the caller before hand.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Link: https://lore.kernel.org/r/20220208031944.3444-2-bjorn.andersson@linaro.org