input/i8042: Add dritek quirk for Acer Aspire One AO532
Like on other Acer devices, the wifi, bluetooth and touchpad on/off toggle
hotkeys on the Acer AO532 do not send any events when the dritek extensions
are not enabled.

Add a quirk to enable the dritek extensions on this netbook model.

Cc: Werner Sembach <wse@tuxedocomputers.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
---
Note this patch applies on top of the:
"[PATCH v5] input/i8042: Rework quirk tables" series
https://lore.kernel.org/linux-input/20220329143622.31394-1-wse@tuxedocomputers.com/