x86/Kconfig: Change depends on X86_PLATFORM_DEVICES to select
The X86_PLATFORM_DEVICES depends on for the Quark and MID options is there
to satisfy the deps of a select on a Kconfig option which itself depends
on X86_PLATFORM_DEVICES.

X86_PLATFORM_DEVICES does not add any kernel code, it merely enables
a submenu. Other Kconfig files which select options from this menu
are already selecting it.

Since mixing select and depends for the some option is generally not
a good idea (it can cause dependency loops), switch over to select.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>