mfd: intel_pmc_bxt: Change depends on X86_PLATFORM_DEVICES to select
The "depends on X86_PLATFORM_DEVICES" for the MFD_INTEL_PMC_BXT option
is there to satisfy the deps of the "select INTEL_SCU_IPC".

X86_PLATFORM_DEVICES does not add any kernel code, it merely enables
a submenu. Other Kconfig files which select options from this menu
are already selecting it.

Since mixing select and depends for the some option is generally not
a good idea (it can cause dependency loops), switch over to select.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>