iio: light: cm32181: Stop suspend failures causing system suspend to fail
On some devices such as the Lenovo ThinkPad X1 Carbon gen 2, the cm32181
might already be powered-down (together with the LCD-panel?) when
the cm32181 suspend callback runs.

This is causing system suspend to fail on these systems.

Instead log a warning on I2C communication errors, while still returning
success from the suspend callback to allow the system suspend to continue.

Link: https://bugzilla.redhat.com/show_bug.cgi?id=2029207
Fixes: 68c1b3dd5c48 ("iio: light: cm32181: Add PM support")
Cc: Kai-Heng Feng <kai.heng.feng@canonical.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>