media: ov2680: Remove ov2680_bayer_order()
The index into ov2680_hv_flip_bayer_order[] shoud be 0-3, but
ov2680_bayer_order() was using 0 + BIT(2) + (BIT(2) << 1) as
max index, which leads to a range of 0-12, so this code can
easily index ov2680_hv_flip_bayer_order[] out of range,
which at best (assuming returning just wrong data) will lead
to a non working setup.

This seems to indicate that the hflip/vlip functionality
was never fully tested.

The ov2680 has a builtin ISP and testing on a device with an atomisp2,
with the atomisp_ov2680 fork of the ov2680.c has shown that with
the default ISP settings for the bayer-order for flipping/mirroring
(which both drivers use) that the bayer order does not change at
all when using flipping/mirroring.

Remove the broken ov2680_bayer_order() function relying on the default
(and correct) MEDIA_BUS_FMT_SBGGR10_1X10 fmt set in ov2680_mode_init().

Signed-off-by: Hans de Goede <hdegoede@redhat.com>