media: atomisp: ov2680: Replace flip ctrls with main ov2680 flip code
Replace the code implementing the flip ctrls with the implementation
from the main drivers/media/i2c/ov2680.c driver.

The goal is to over time make atomisp-ov2680.c more (and more)
like the standard ov2680.c driver so that it can eventually be
replaced with the standard ov2680.c driver.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>