platform: x86: int3472: Register a LED lookup table entry for the privacy LED
All currently known models using the tps68470 PMIC have a privacy LED for
the back sensor (first ACPI consumer dev of the PMIC) connected to the
ileda output of the PMIC.

Add a LED lookup table entry for this, so that the v4l2-core code turns on
the LED when streaming from the sensor.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
---
Kate, please include this in the v2 series adding the leds-tps68470 driver.