media: atomisp: ov2680: Convert to new CCI register access helpers
Use the new comon CCI register access helpers to replace the private
register access helpers in the ov2680 driver.

While at it also switch to using the same register address defines
as the standard drivers/media/i2c/ov2680.c driver to make merging
the 2 drivers simpler.

Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
---
Note atomisp-ov2680.c is going away. The main atomisp driver now
supports standard v4l2 drivers and a patch series is pending for
the main drivers/media/i2c/ov2680.c to get it into shape to
replace atomisp-ov2680.c .

The only reason to upstream this patch is because much
of the work landing in the main ov2680.c is copy -pasted
from the state of atomisp-ov2680.c *after this patch* ,
so having this in git history before deleting atomisp-ov2680.c
is helpful in case someone ever finds the need to compare
the code.