media: dt-bindings: media: i2c: GC5035: Fix GPIOs to use standardizes con-ids
Some boards have standardized sensor connectors which allow
connecting various sensor modules.

On x86/ACPI devices the routing of these connectors is described
with a custom ACPI mechanism. The kernel's int3472 driver maps
the custom ACPI mechanism to standard kernel resources.

This mapping uses standardized con-ids for the various resources,
specifically it uses "powerdown" for the powerdown / shutdown signal
and "reset" for the reset signal.

This commit updates the dt-bindings for the GC5035 to use these standard
con-ids so that there is interoperability between DT and ACPI enumerated
fwnodes / resources for the GC5035.

Since this is part of the same patch-series as merging the initial
dt-bindings this is not a DT ABI break from the view point of
previously documented DT bindings.

However I assume that this driver which was originally submitted
by ChromeOS devs but never mainlined is already in use on ChromeOS.

So this patch assumes that for ChromeOS Google will be able to
update the kernel (switching to the mainlined driver) and the DTB
file in sync. If that is not possible this patch should probably
be dropped (fixing the ACPI interop in another way).

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
---
Note that for "resetb" vs "reset"  the original v4 submission
already contained some confusion. The original bindings specified
"resetb-gpios" but the provided example used "reset-gpios"