media: gc5035: Drop voltage suffix from regulator supply-ids
The voltage of the regulator is a separate property which should
not be encoded in the supply-id, especially since different designs
may use somewhat different voltages.

Dropping the voltage suffix for avdd and dvdd also makes their
supply-ids consistent with iovdd which already did not have
a voltage suffix.

Dropping the voltage suffix from the avdd-supply id also fixes
a mismatch between the bindings and the driver in the original
upstream v4 submission.

In the v4 patch-set originally submitted upstream the dt-bindings were
using a "28" suffix ("avdd28"), where as the driver was using "avdd21"
when trying to get the regulator.

Dropping the suffix also provides interoperability between DT and
ACPI enumerated fwnodes / resources for the GC5035. On x86/ACPI devices
the routing of the camera connectors is described with a custom ACPI
mechanism. The kernel's int3472 driver maps this custom ACPI mechanism
to standard kernel resources using "avdd" for the regulator supply-id.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>