media: gc5035: Fix GPIO lines con-ids
Some boards have standardized sensor connectors which allow
connecting various sensor modules.

On x86/ACPI devices the routing of these connectors is described
with a custom ACPI mechanism. The kernel's int3472 driver maps
the custom ACPI mechanism to standard kernel resources.

This mapping uses standardized con-ids for the various resources,
specifically it uses "powerdown" for the powerdown / shutdown signal
and "reset" for the reset signal.

Change the driver to use the standard "powerdown" and "reset"
con-ids when getting the GPIOs for interoperability with
ACPI enumerated fwnodes / resources.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>