media: atomisp: csi2-bridge: Use dev_info() instead of acpi_handle_info()
acpi_handle_info() uses the ACPI path to the handle as prefix for messages
which often is not really informative, e.g. it shows both sensors as
_SB_.I2C2.CAM8 / .CAM2 rather then using OVTI2680 / INT0310 as prefix
making it unnecessary diffucult to figure out which message belongs to
which sensor.

This is double confusing because later messages from the actual sensor
drivers do use OVTI2680 / INT0310 as prefix.

Switch to dev_info() / dev_warn() / dev_err() to consistently use
e.g. OVTI2680 / INT0310 as prefix.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>